#956 Phase 5: extract cos/queue_ops.rs from tx.rs
Per docs/pr/956-phase5-queue-ops/plan.md (v7, Codex 7 rounds +
Gemini both PLAN-READY), extract the queue-state lifecycle
subsystem from tx.rs into cos/queue_ops.rs.

Move list (~787 LOC removed from tx.rs):

Functions (18 total):
- 14 always-on pub(in crate::afxdp) re-exports:
  cos_item_len, cos_queue_clear_orphan_snapshot_after_drop,
  cos_queue_drain_all, cos_queue_front, cos_queue_is_empty,
  cos_queue_len, cos_queue_pop_front, cos_queue_pop_front_no_snapshot,
  cos_queue_push_back, cos_queue_push_front, cos_queue_restore_front,
  cos_queue_v_min_consume_suspension, cos_queue_v_min_continue,
  publish_committed_queue_vtime
- 2 file-private helpers:
  cos_queue_pop_front_inner, cos_queue_min_finish_bucket
- 2 cfg-gated test-only re-exports:
  account_cos_queue_flow_enqueue, account_cos_queue_flow_dequeue

Constants (5):
- 3 file-private: V_MIN_READ_CADENCE, V_MIN_LAG_THRESHOLD_NS,
  V_MIN_MIN_LAG_BYTES
- 2 cfg-gated test-only re-exports: V_MIN_CONSECUTIVE_SKIP_HARD_CAP,
  V_MIN_SUSPENSION_BATCHES

Helper: compute_v_min_lag_threshold (file-private — only caller is
cos_queue_v_min_continue).

#[inline] applied on per-byte hot-path fns (push_back, push_front,
pop_front, pop_front_inner, account_*, cos_item_len, etc.) per the
Phase 4 lesson — pub(in crate::afxdp) plus #[inline] preserves
cross-module inlining in release builds.

Resolves Phase 3 deferral: Gemini round-1 of Phase 3 deferred
account_* helpers to Phase 5 because MQFQ + V-min invariants must
stay cohesive with selection/pop/publish. This PR delivers that
cohesion — the entire MQFQ ordering bookkeeping + V-min slot
lifecycle now lives in one file.

CoSBatch / CoSServicePhase / ExactCoSQueueKind enums and their
consumers (select_cos_*_batch, service_exact_*_queue_direct) stay
in tx.rs through Phase 7 (queue_service) — they live with the
dispatch entry points, not the queue state primitives.

worker.rs adds explicit `use super::cos::{cos_queue_len,
cos_queue_pop_front_no_snapshot, ...}` because the previous tx-glob
path is gone.

cos/admission.rs header note past-tense'd: account_* now live in
cos/queue_ops.rs.

Net change: -787 lines from tx.rs (17030 -> 16245).

Tests: cargo build --bins clean (no Phase-5 unused-import
warnings); cargo test --bins -> 865 passed, 0 failed, 2 ignored
(same baseline as Phase 4).

Reviewers signed off:
- Codex rounds 1-7 (gpt-5.5 xhigh, PLAN-READY at v7)
- Gemini round-1 (PLAN-READY at v4 on systems-level concerns:
  hot-path correctness, MQFQ + V-min cohesion, atomic ordering
  preserved)
